Add support for TLS encryption
Certificates and keys are now accepted as command line options, and when
given will force the client to negotiate a TLS connection with the
server.  It is a fatal error if negotiating TLS fails.  We do not fall
back to an insecure connection.

Hostname validation code is from /usr/src/contrib/libevent/sample/.