Don't chdir or redirect to /dev/null when daemonizing
We have to call daemon(3) before connecting otherwise we lose the
socket.  So in order for relative certificate lookups and message output
(such as the name of the ggate device created) to work, don't chdir to /
or redirect stdin/stdout/stderr to /dev/null.

There may be a better way to fix this, since the later call to daemon(3)
worked fine before switching from the system.dns to system.net Casper
service.