Close stdout/stderr after printing the device name
As a consequence of calling daemon(3) earlier and keeping stdout/stderr
open, we now have to manually close the files after completing the setup
and printing the device name.  This ensures a pipeline such as
`nbd-client localhost | tee foo` does not leave tee waiting indefinitely
for its stdin to close.