ci: Fix diskinfo aio error on STABLE branches
FreeBSD stable branches use the latest package repos which have a newer
version of nbdkit that advertises 32k blocksize for the memory backend
instead of 512.  This causes diskinfo to fail when it tries doing 4k
async random reads because that is smaller than the sectorsize.

There is no way to configure the server to prefer 4k blocksize with the
older version of nbdkit in the quarterly package repos used by release
branches, so force the client to use 4k sectors.